requirements.txt
aiohttp>=3.6.0,<4
discord.py[speed] @ git+https://github.com/Rapptz/discord.py
jishaku
json5
lru-dict
motor
pymongo[srv]
python-dotenv
uvloop
sentry-sdk